test: add more wait in create_cmd, load ntdll then check in get_module_in_pid
The wait in create_cmd was sometimes not enough for Github actions.
So
I
increased the wait to 50ms, after process creation. That should
help
for
windows to get to the initialisation, before I read stuff

In
the
test
for get_module_in_pid, I now load ntdll.dll, before I check for
it.